feat(backend): Code refactoring: - Extracting some private util methods as extension functions - Convert TimeRangeSplitter into a singleton object - Fix tests with wrong mocking data

+++ backend/src/main/kotlin/metrik/metrics/domain/model/CalculationPeriod.kt
@@ -1,5 +1,5 @@
 package metrik.metrics.domain.model
 
-enum class CalculationPeriod(val timeInDays: Int){
-    Fortnightly(14), Monthly(30)
+enum class CalculationPeriod(val timeInDays: Long){
+    Fortnightly(14L), Monthly(30L)
 }
